+++ src/routes/admin/listas/financeiro/+page.svelte
@@ -26,8 +26,7 @@
 	<div class="flex items-center justify-between">
 		<h1 class="text-2xl font-bold">Usuários do Financeiro</h1>
 		<div class="flex gap-2">
-			<Input placeholder="Buscar por nome..." bind:value={searchTerm} class="w-64" />
-			<Button variant="outline">Adicionar Usuário</Button>
+			<Input placeholder="Buscar por nome..." bind:value={searchTerm} class="w-96" />
 		</div>
 	</div>
