+++ src/routes/api/admin/listar/vendedores-internos/+server.ts
@@ -3,7 +3,7 @@ import { userTable } from '$lib/server/database/schema';
 import type { RequestHandler } from './$types';
 import { db } from '$lib/server/database/db.server';
 import { eq } from 'drizzle-orm';
-import type { PageData } from '../../../admin/listas/administradores/types';
+import type { VendedoresResponse } from '../../../../admin/listas/vendedores-internos/types';
 
 const validateApiKey = (request: Request): boolean => {
 	return request.headers.get('API-KEY') === SITE_CHAVE_API;
@@ -12,20 +12,20 @@ const validateApiKey = (request: Request): boolean => {
 export const GET: RequestHandler = async ({ request, locals }) => {
 	try {
 		if (!validateApiKey(request)) {
-			return new Response(JSON.stringify({ error: 'Chave de API inválida', administradores: [] }), {
+			return new Response(JSON.stringify({ error: 'Chave de API inválida', vendedores: [] }), {
 				status: 401,
 				headers: { 'Content-Type': 'application/json' }
 			});
 		}
 
 		if (!locals.user || locals.user.job !== 'Admin') {
-			return new Response(JSON.stringify({ error: 'Não autorizado', administradores: [] }), {
+			return new Response(JSON.stringify({ error: 'Não autorizado', vendedores: [] }), {
 				status: 401,
 				headers: { 'Content-Type': 'application/json' }
 			});
 		}
 
-		const administradores = await db
+		const vendedoresInternos = await db
 			.select({
 				id: userTable.id,
 				name: userTable.name,
@@ -35,10 +35,10 @@ export const GET: RequestHandler = async ({ request, locals }) => {
 				criadoEm: userTable.criadoEm
 			})
 			.from(userTable)
-			.where(eq(userTable.job, 'Admin'));
+			.where(eq(userTable.job, 'Vendedor Interno'));
 
-		const response: PageData = {
-			administradores,
+		const response: VendedoresResponse = {
+			vendedores: vendedoresInternos,
 			error: null
 		};
 
@@ -46,12 +46,12 @@ export const GET: RequestHandler = async ({ request, locals }) => {
 			headers: { 'Content-Type': 'application/json' }
 		});
 	} catch (error) {
-		console.error('Erro ao buscar administradores:', error);
+		console.error('Erro ao buscar vendedores internos:', error);
 		return new Response(
 			JSON.stringify({
 				error: 'Erro interno do servidor',
-				administradores: []
-			} satisfies PageData),
+				vendedores: []
+			} satisfies VendedoresResponse),
 			{
 				status: 500,
 				headers: { 'Content-Type': 'application/json' }
